backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
mongoengine==0.29.1
PyJWT>=2.8
django-cors-headers>=4.0
Pillow>=10.0
python-dotenv>=1.1
werkzeug>=3.0
